Allow configuration of kotlinx serialization
